back/ecommerce-fabrizio-zeballos/src/modules/categories/categories.controller.ts
import { Controller, Get } from '@nestjs/common';
import { CategoriesService } from './categories.service';
import { Category } from '../../entities/categories.entity';
import { ApiTags } from '@nestjs/swagger';

@Controller('categories')
@ApiTags('Categories')
export class CategoriesController {
  constructor(private readonly categoriesService: CategoriesService) {}

  @Get()
  async getCategories(): Promise<Category[]> {
    return await this.categoriesService.getCategories();
  }

  @Get('seeder')
  async addCategories(): Promise<string> {
    return await this.categoriesService.addCategories();
  }
}
